Update SQL_Supply_Builder.py
Updated program, added loop to scan local dir for an existing db, if no db found asks for a location or to create a new one. At this time only handles one DB, unable to select multiple DB from a list. 

Still in progress: creating a .csv or a .xml file to read from where the db is so scanning a directory or constantly asking for location is necessary.